+++ src/common/hooks/index.js
@@ -1,8 +1,10 @@
+export { default as useBoundPieces } from './useBoundPieces';
 export { default as useControlledAccordion } from './useControlledAccordion';
 export { default as useInstance } from './useInstance';
 export { default as useGoBack } from './useGoBack';
 export { default as useHolding } from './useHolding';
 export { default as useInstanceQuery } from './useInstanceQuery';
+export { default as usePiecesMutation } from './usePiecesMutation';
 export { default as useSearchInstanceByIdQuery } from './useSearchInstanceByIdQuery';
 export { default as useTenantKy } from './useTenantKy';
 export { default as useDefaultJobProfile } from './useDefaultJobProfile';